requirements.txt
absl-py==0.7.0
astor==0.7.1
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
imutils==0.5.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.0.1
mock==2.0.0
more-itertools==6.0.0
numpy==1.16.2
pbr==5.1.3
Pillow==5.4.1
protobuf==3.7.0
PyYAML==3.13
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==0.14.1
